+++ packages/patrol_finders/lib/src/custom_finders/patrol_finder.dart
@@ -403,9 +403,12 @@ class PatrolFinder implements MatchFinder {
   ///
   /// Timeout is globally set by [PatrolTester.config.visibleTimeout]. If you
   /// want to override this global setting, set [timeout].
+  ///
+  /// {@macro patrol_tester.alignment_on_visible_check}
   Future<PatrolFinder> waitUntilVisible({
     Duration? timeout,
     bool enablePatrolLog = true,
+    Alignment alignment = Alignment.center,
   }) =>
       wrapWithPatrolLog(
         action: 'waitUntilVisible',
@@ -414,6 +417,7 @@ class PatrolFinder implements MatchFinder {
           this,
           timeout: timeout,
           enablePatrolLog: false,
+          alignment: alignment,
         ),
         enablePatrolLog: enablePatrolLog,
       );
@@ -508,9 +512,12 @@ class PatrolFinder implements MatchFinder {
   @override
   String describeMatch(Plurality plurality) => finder.describeMatch(plurality);
 
-  /// Returns true if this finder finds at least 1 visible widget.
-  bool get visible {
-    final isVisible = hitTestable().evaluate().isNotEmpty;
+  /// Returns true if this finder finds at least 1 visible widget
+  /// at the given [alignment].
+  ///
+  /// {@macro patrol_tester.alignment_on_visible_check}
+  bool isVisibleAt({Alignment alignment = Alignment.center}) {
+    final isVisible = hitTestable(at: alignment).evaluate().isNotEmpty;
     if (isVisible == true) {
       assert(
         exists == true,
@@ -520,6 +527,14 @@ class PatrolFinder implements MatchFinder {
 
     return isVisible;
   }
+
+  /// Returns true if this finder finds at least 1 visible widget.
+  ///
+  /// will call [isVisibleAt] with [Alignment.center]
+  ///
+  /// In case this returns false and you are sure that the widget is visible,
+  /// try calling [isVisibleAt] with a different [Alignment] parameter.
+  bool get visible => isVisibleAt();
 
   @override
   FinderResult<Element> evaluate() => finder.evaluate();