+++ packages/patrol_finders/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Add `isVisibleAt` method to check if a widget is visible at a given alignment in case `visible` fails.
 - Remove `exception` from `StepEntry`. When it was too long, it caused crash because of badly formed JSON
 - Bump `patrol_log` version.
+- Bump compileSdk to 35.
+- Bump gradle to 8.9.0.
 
 ## 2.6.0